+++ src/GSvar/MainWindow.cpp
@@ -13,6 +13,7 @@
 #include <QDesktopServices>
 #include <QUrl>
 #include <QTime>
+#include <QNetworkProxy>
 #include "ExternalToolDialog.h"
 #include "ReportDialog.h"
 #include <QInputDialog>
@@ -3169,8 +3170,10 @@ void MainWindow::on_actionAbout_triggered()
 			add_info += "<br>&nbsp;&nbsp;start time: " + server_info.server_start_time.toString("yyyy-MM-dd hh:mm:ss");
 			add_info += "<br>&nbsp;&nbsp;API URL: " + server_info.server_url;
 			add_info += "<br>&nbsp;&nbsp;API version: " + server_info.api_version;
-			add_info += "<br>&nbsp;&nbsp;htslib version: " + server_info.htslib_version;
+			add_info += "<br>&nbsp;&nbsp;operating system: " + server_info.operating_system;
+			add_info += "<br>&nbsp;&nbsp;architecture: " + server_info.architecture;
 			add_info += "<br>&nbsp;&nbsp;Qt version: " + server_info.qt_version;
+			add_info += "<br>&nbsp;&nbsp;htslib version: " + server_info.htslib_version;
 		}
 	}
 
@@ -3954,13 +3957,6 @@ void MainWindow::generateReportSomaticRTF()
 			report.storeRtf(temp_filename);
 			Helper::moveFile(temp_filename, file_rep);
 			Log::perf("Generating somatic report RTF took ", timer);
-
-			//Generate files for QBIC upload
-			timer.start();
-			QString path = ps_tumor + "-" + ps_normal;
-			if (GlobalServiceProvider::fileLocationProvider().isLocal()) path = Settings::string("qbic_data_path") + "/" + path;
-			report.storeQbicData(path);
-			Log::perf("Generating somatic report QBIC data took ", timer);
 
 			QApplication::restoreOverrideCursor();
 		}
@@ -5001,6 +4997,20 @@ void MainWindow::on_actionNotifyUsers_triggered()
 
 	EmailDialog dlg(this, to, subject, body);
 	dlg.exec();
+}
+
+void MainWindow::on_actionDetermineProxy_triggered()
+{
+	QString url = QInputDialog::getText(this, "Determine proxy for URL", "URL:");
+	if (url.isEmpty()) return;
+
+	QStringList proxies;
+	foreach(const QNetworkProxy& proxy, QNetworkProxyFactory::systemProxyForQuery(QNetworkProxyQuery(url)))
+	if (proxy.type()!=QNetworkProxy::NoProxy)
+	{
+		proxies << (proxy.hostName() + ":" + QString::number(proxy.port()));
+	}
+	QMessageBox::information(this, "Determine proxy for URL", "For the URL '" + url + "' the follwing proxies are used:\n" + proxies.join('\n'));
 }